orbital-os/scripts/orb_metrics_requirements.txt
numpy>=1.24
Pillow>=10.0
scikit-image>=0.22
torch>=2.0
lpips>=0.1.4
